pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "longxiaclaw"
version = "0.1.0"
description = "An OAuth-agnostic personal AI agent. Designed for lightweight, LLM CLI-powered, skill-extensible operation."
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "croniter>=2.0",
    "python-dotenv>=1.0",
    "ddgs>=9.0",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
longxia = "longxiaclaw.main:cli"

[tool.setuptools]
packages = [
    "longxiaclaw",
    "longxiaclaw.backends",
    "longxiaclaw.channels",
    "longxiaclaw.memory",
    "longxiaclaw.scheduler",
    "longxiaclaw.skills",
    "longxiaclaw.system",
    "longxiaclaw.tools",
    "longxiaclaw.tui",
]

[tool.setuptools.package-dir]
longxiaclaw = "longxiaclaw"

[tool.pytest.ini_options]
testpaths = ["longxiaclaw/tests"]
asyncio_mode = "auto"
markers = [
    "network: tests that require network access",
]

[tool.ruff]
target-version = "py310"
line-length = 100
